website: move video embed to own file
